protos: Collapse ClearMessage template
We introduce Ptr removal so that our helper functions can be agnostically passed in a Ptr<T> or a T*.

We utilize SFINAE to determine if PtrOrRaw. If the target is const, or not in {T, T*, Ptr<T>}, the template will not match and be discarded (but still compile successfully).

PiperOrigin-RevId: 636688950